+++ moor/lib/src/runtime/query_builder/query_builder.dart
@@ -26,7 +26,7 @@ part 'expressions/datetimes.dart';
 part 'expressions/expression.dart';
 part 'expressions/in.dart';
 part 'expressions/null_check.dart';
-part 'expressions/subquery.dart';
+part 'expressions/exists.dart';
 part 'expressions/text.dart';
 part 'expressions/variables.dart';
